Move PresidioApiHandler, split interface from implementation Fix Tests CRLF
