+++ src/Microsoft.Health.Fhir.Anonymizer.Shared.Core.UnitTests/Microsoft.Health.Fhir.Anonymizer.Shared.Core.UnitTests.projitems
@@ -14,6 +14,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)AnonymizerConfigurations\AnonymizerConfigurationTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)AnonymizerConfigurations\AnonymizerConfigurationValidatorTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)AnonymizerEngineTests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Api\PresidioApiHandlerMock.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\ElementNodeNavExtensionTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\ElementNodeVisitorExtensionsTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\ElementTypeExtensionsTests.cs" />
@@ -28,7 +29,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)Processors\EncryptProcessorTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Processors\GeneralizeProcessorTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Processors\PerturbProcessorTests.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Processors\PresidioApiHandlerMock.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Processors\PresidioProcessorTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Processors\RedactProcessorTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Processors\Settings\GeneralizeSettingTests.cs" />
