+++ src/Microsoft.Health.Fhir.Anonymizer.Shared.Core.UnitTests/Processors/PresidioProcessorTests.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Fhir.Anonymizer.Core.UnitTests.Processors;
+﻿using Fhir.Anonymizer.Core.UnitTests.Api;
 using Hl7.Fhir.ElementModel;
 using Hl7.Fhir.Model;
 using Microsoft.Health.Fhir.Anonymizer.Core.Processors;
