+++ src/Microsoft.Health.Fhir.Anonymizer.Shared.Core/Api/ApiHandler.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+using Presidio.Model;
+
+namespace Microsoft.Health.Fhir.Anonymizer.Core.Api
+{
+    public interface ApiHandler
+    {
+        List<RecognizerResult> Analyze(string text);
+        AnonymizeResponse Anonymize(string text, List<RecognizerResult> analyzerResult);
+    }
+}
