+++ src/Microsoft.Health.Fhir.Anonymizer.Shared.Core/Utility/PresidioUtility.cs
@@ -1,10 +1,12 @@
 ﻿
+using Microsoft.Health.Fhir.Anonymizer.Core.Api;
+
 namespace Microsoft.Health.Fhir.Anonymizer.Core.Utility
 {
     public class PresidioUtility
 
     {        
-        public static string Anonymize(string text, IApiHandler apiHandler)
+        public static string Anonymize(string text, ApiHandler apiHandler)
         {
             var analyzerResult = apiHandler.Analyze(text);
             var anonymizerResponse = apiHandler.Anonymize(text, analyzerResult);
